+++ src/bl64-cnt-setup.bash
@@ -63,7 +63,7 @@ function _bl64_cnt_set_command() {
   elif [[ -x "$BL64_CNT_CMD_PODMAN" ]]; then
     BL64_CNT_DRIVER="$BL64_CNT_DRIVER_PODMAN"
   else
-    bl64_msg_show_error "unable to find a container manager (${BL64_CNT_CMD_DOCKER}, ${BL64_CNT_CMD_PODMAN})"
+    bl64_msg_show_error "unable to find a container manager CLI location (${BL64_CNT_CMD_DOCKER}, ${BL64_CNT_CMD_PODMAN})"
     return $BL64_LIB_ERROR_APP_MISSING
   fi
   bl64_dbg_lib_show_vars 'BL64_CNT_DRIVER'
@@ -88,10 +88,11 @@ function _bl64_cnt_set_command_docker() {
       # Rancher Desktop using docker
       command_location="${HOME}/.rd/bin"
     fi
+  else
+    bl64_check_directory "$command_location" || return $?
   fi
 
-  bl64_check_directory "$command_location" || return $?
-  [[ -x "${command_location}/docker" ]] && BL64_CNT_CMD_DOCKER="${command_location}/docker"
+  BL64_CNT_CMD_DOCKER="${command_location}/docker"
   return 0
 }
 
@@ -109,10 +110,11 @@ function _bl64_cnt_set_command_podman() {
     elif [[ -x '/usr/bin/podman' ]]; then
       command_location='/usr/bin'
     fi
+  else
+    bl64_check_directory "$command_location" || return $?
   fi
 
-  bl64_check_directory "$command_location" || return $?
-  [[ -x "${command_location}/podman" ]] && BL64_CNT_CMD_PODMAN="${command_location}/podman"
+  BL64_CNT_CMD_PODMAN="${command_location}/podman"
   return 0
 }