+++ src/bl64-iam-setup.bash
@@ -141,6 +141,7 @@ function _bl64_iam_set_alias() {
 function _bl64_iam_set_options() {
   bl64_dbg_lib_show_function
 
+  # shellcheck disable=SC2034
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_KL}-*)
     BL64_IAM_SET_USERADD_CREATE_HOME='--create-home'