+++ src/bl64-msg-setup.bash
@@ -173,10 +173,10 @@ function bl64_msg_set_output() {
   local theme="${2:-${BL64_VAR_DEFAULT}}"
 
   if bl64_lib_var_is_default "$output"; then
-    if [[ "$-" == *i* ]]; then
-      output="$BL64_MSG_OUTPUT_ANSI"
-    else
+    if bl64_lib_flag_is_enabled "$BL64_LIB_CICD"; then
       output="$BL64_MSG_OUTPUT_ASCII"
+    else
+      output="$BL64_MSG_OUTPUT_ANSI"
     fi
   fi